classify test case and add test case for corrupt file
