+++ src/test/java/com/github/junrar/JunrarTest.java
@@ -3,17 +3,17 @@ package com.github.junrar;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 
 import com.github.junrar.vfs2.provider.rar.FileSystem;
 import org.apache.commons.io.FileUtils;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import com.github.junrar.exception.RarException;
@@ -24,12 +24,12 @@ public class JunrarTest {
     private static File tempFolder;
     private FileSystem fileSystem = new FileSystem();
 
-    @BeforeClass
+    @Before
     public static void setupFunctionalTests() throws IOException {
         tempFolder = TestCommons.createTempDir();
     }
 
-    @AfterClass
+    @After
     public static void tearDownFunctionalTests() throws IOException {
         FileUtils.deleteDirectory(tempFolder);
     }
@@ -99,26 +99,23 @@ public class JunrarTest {
     }
 
     @Test
-    public void ifIsDirInsteadOfFile_ThrowException() throws RarException, IOException {
+    public void extractRarV4() throws Exception {
+        InputStream stream = null;
         try {
-            Junrar.extract(tempFolder, tempFolder);
-        } catch (final IllegalArgumentException e) {
-            assertEquals("First argument should be a file but was " + tempFolder.getAbsolutePath(), e.getMessage());
-            return;
+            final File rarV4 = TestCommons.writeResourceToFolder(tempFolder, "rar4.rar");
+            stream = new FileInputStream(rarV4);
+            Junrar.extract(stream, tempFolder);
+        } finally {
+            if (stream != null) {
+                stream.close();
+            }
         }
-        fail();
-    }
-
-    @Test
-    public void ifIsFileInsteadOfDir_ThrowException() throws RarException, IOException {
-        final File rarFileOnTemp = TestCommons.writeTestRarToFolder(tempFolder);
-        try {
-            Junrar.extract(rarFileOnTemp, rarFileOnTemp);
-        } catch (final IllegalArgumentException e) {
-            assertEquals("the destination must exist and point to a directory: " + rarFileOnTemp.getAbsolutePath(), e.getMessage());
-            return;
-        }
-        fail();
+        final File file1 = new File(tempFolder, "FILE1.TXT");
+        final File file2 = new File(tempFolder, "FILE2.TXT");
+        assertTrue(file1.exists());
+        assertEquals(7, file1.length());
+        assertTrue(file2.exists());
+        assertEquals(7, file2.length());
     }
 
     private static ContentDescription c(final String name, final long size) {
